합쳐서 기능 한꺼번에 쓸 수 있도록 수정 #176
